update t0 3.2.9
git-svn-id: https://plugins.svn.wordpress.org/super-static-cache/trunk@1269078 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ super-static-cache.php
@@ -3,7 +3,7 @@
 Plugin Name: Super Static Cache
 Plugin URI: https://www.hitoy.org/super-static-cache-for-wordperss.html
 Description: Super Static Cache is an efficient WordPress caching engine which provides three cache mode. It can reduce the pressure of the database significantly that makes your website faster than ever.
-Version: 3.2.8
+Version: 3.2.9
 Author: Hito
 Author URI: https://www.hitoy.org/
 Text Domain: super_static_cache
@@ -110,20 +110,20 @@ function mkdirs($path){
 		return true;
 }
 //给目录赋予权限
-function chmods($path,$mod=0777,$rec=true){
-		if($rec==false){
-				return chmod($path,$mod);
+function chmods($path,$dirmod=0777,$filemod=0666,$rec=true){
+		if($rec==false || is_file($path)){
+				return @chmod($path,$filemod);
 		}
 		if(is_dir($path)){
-				chmod($path,$mod);
+				@chmod($path,$dirmod);
 				$dir=opendir($path);
 				while(($file = readdir($dir)) !== false){
 						if($file == '.' || $file == '..') continue;
 						$ffile=$path.'/'.$file;
-						if(is_file($ffile))  chmod($ffile,$mod);
+						if(is_file($ffile))  @chmod($ffile,$filemod);
 						if(is_dir($ffile)) {
-								chmod($ffile,$mod);
-								if($rec==true) chmods($ffile,$mod,$rec);
+								@chmod($ffile,$dirmod);
+								if($rec==true) chmods($ffile,$dirmod,$filemod,$rec);
 						}
 				}
 				closedir($dir);
@@ -293,17 +293,17 @@ class WPStaticCache{
 				if(!$this->is_pagetype_support_cache()) return false;
 
 				//对含有查询的情况进行过滤
-				preg_match("/^([^?]+)?/i",$this->wpuri,$match);
-				$realname=urldecode($match[1]);
+				preg_match("/^\/([^?]+)?/i",$this->wpuri,$match);
+				$realname=!empty($match[1])?urldecode($match[1]):"";
+
 				//去掉目录之后的文件名
 				$fname=substr($realname,strripos($realname,"/")+1);
-
+					
 				if($this->cachemod == 'serverrewrite' || $this->cachemod == 'phprewrite'){
-						$cachedir='super-static-cache';
-				}else {
+						$cachedir='super-static-cache/';
+				}else if($this->cachemod == 'direct'){
 						$cachedir='';
 				}
-
 				if($fname == ""){
 						//以'/'结尾的请求
 						$cachename = $this->wppath.$cachedir.$realname."index.html";
@@ -322,27 +322,31 @@ class WPStaticCache{
 				return $cachename;
 		}
 
-		//写入并保存缓存，最终动作
+		//写入缓存，并赋予相应的文件权限便于其它工具进行处理
 		//满足三种情况
 		//1, url能缓存 filename存在
-		//2, 文件名不存在(保护原有的文件不被改写)
-		//3, 缓存的内容不为空
+		//2, 缓存的内容不为空
+		//3, 文件名不存在(保护原有的文件不被改写)
 		public function save_cache_content(){
 				$filename = $this->get_cache_fname();
-				if($filename && !file_exists($filename) && strlen($this->htmlcontent) > 0){
+				if($filename && strlen($this->htmlcontent) > 0 && !file_exists($filename)){
 						//创建存放缓存的目录
 						mkdirs(dirname($filename));
 						//加锁写入缓存
 						file_put_contents($filename,$this->htmlcontent,LOCK_EX);
 
-						//给于缓存的操作权限
-						if($this->cachemod == 'serverrewrite' || $this->cachemod == 'phprewrite'){
-								$cachedir='super-static-cache';
-						}else {
-								$cachedir='';
+						//对缓存文件的权限进行更改
+						$relauri=substr($filename,strlen($this->wppath)-1);
+						preg_match("/^\/([^?]+)?/i",$relauri,$match);
+						$realname=!empty($match[1])?urldecode($match[1]):"";
+						$relapath=substr($realname,0,strpos($realname,'/'));
+						if($relapath==""){
+							chmods($filename,0777,0666,false);
+						}else if($relapath != "" && $this->cachemod == "direct"){
+							chmods($this->wppath.$relapath,0777,0666,true);
+						}else if($relapath != "" && $this->cachemod == "serverrewrite" || $this->cachemod == "phprewrite"){
+							chmods($this->wppath."super-static-cache/".$relapath,0777,0666,true);
 						}
-						$modir=substr($filename,0,strlen($this->wppath.$cachedir));
-						//chmods($modir,0777);
 				}
 		}
 
@@ -405,7 +409,7 @@ class WPStaticCache{
 		//安装函数
 		public function install(){
 				add_option("super_static_cache_mode","close");
-				add_option("super_static_cache_strict",false);
+				add_option("super_static_cache_strict",true);
 				add_option("super_static_cache_excet","author,date,attachment");
 				add_option("update_cache_action","publish_post,post_updated,trashed_post,publish_page");
 
@@ -414,7 +418,7 @@ class WPStaticCache{
 						mkdir($this->wppath.'super-static-cache',0777);
 				}
 				file_put_contents($this->wppath."super-static-cache/rewrite_ok.txt","This is a test file from rewrite rules,please do not to remove it.\n");
-				chmods($this->wppath.'super-static-cache',0777);
+				chmods($this->wppath.'super-static-cache',0777,0444,true);
 		}
 		//卸载函数
 		public function unistall(){
@@ -422,11 +426,11 @@ class WPStaticCache{
 				delete_option("super_static_cache_excet");
 				delete_option("super_static_cache_strict");
 				delete_option("update_cache_action");
-				//删除
+				//删除一些必要的缓存
 				delete_uri($this->wppath."super-static-cache/rewrite_ok.txt");
 				delete_uri($this->wppath.'super-static-cache');
-				if($this->cachemod=='direct'){
-						delete_uri($this->wppath.'index.html');
+				if($this->cachemod=="direct" && is_file($this->wppath."index.html")){
+						delete_uri($this->wppath."index.html");
 				}
 		}
